refactor(ListItemButtonIcon): remove confirm property
