+++ packages/renderer/src/lib/volume/VolumeActions.svelte
@@ -2,6 +2,8 @@
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
 import { createEventDispatcher } from 'svelte';
 
+import { withConfirmation } from '/@/lib/dialogs/MessageBoxUtils';
+
 import ListItemButtonIcon from '../ui/ListItemButtonIcon.svelte';
 import type { VolumeInfoUI } from './VolumeInfoUI';
 
@@ -21,8 +23,7 @@ async function removeVolume(): Promise<void> {
 {#if volume.status === 'UNUSED'}
   <ListItemButtonIcon
     title="Delete Volume"
-    confirm="{true}"
-    onClick="{() => removeVolume()}"
+    onClick="{() => withConfirmation(removeVolume, 'Delete Volume')}"
     detailed="{detailed}"
     icon="{faTrash}" />
 {/if}
